pyproject.toml
[project]
name = "narrafind"
version = "0.1.0"
description = "Search inside long videos using natural language — find what was shown or said"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "python-dotenv>=1.0",
    "chromadb>=0.4",
    "google-genai>=1.0",
    "imageio-ffmpeg>=0.5",
    "flask>=3.0",
    "flask-cors>=4.0",
    "openai-whisper",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
narrafind = "narrafind.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["narrafind"]

[dependency-groups]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["narrafind"]

[tool.coverage.report]
show_missing = true
